% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.character.PJRTElementType}
\alias{as.character.PJRTElementType}
\title{Convert \code{PJRTElementType} to string}
\usage{
\method{as.character}{PJRTElementType}(x, ...)
}
\arguments{
\item{x}{A PJRT element type object.}

\item{...}{Additional arguments (unused).}
}
\value{
A string representation of the element type.
}
\description{
Get a (lowercase) string representation of a PJRT element type
}
