% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as_array}
\alias{as_array}
\title{Convert a PJRT Buffer to an R Array}
\usage{
as_array(buffer, client = default_client())
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
A PJRT buffer object.}

\item{client}{(\code{PJRTClient})\cr
A PJRT client object.}
}
\description{
Copy a \code{\link[=pjrt_buffer]{PJRTBuffer}} to an R array.
}
\details{
Currently, two copies are being done:
\enumerate{
\item Copy the data from the PJRT device to the CPU.
\item Copy the data from the CPU to the R array.
}
}
