% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{as_pjrt_client}
\alias{as_pjrt_client}
\title{Convert to PJRT Client}
\usage{
as_pjrt_client(x)
}
\arguments{
\item{x}{(\code{PJRTClient} | \code{character(1)})\cr
Either a PJRT client object or a platform name (e.g., "cpu", "cuda").}
}
\value{
\code{PJRTClient}
}
\description{
Convert a platform name to a PJRT client or verify that an object is already a client.
}
\examples{
\dontshow{if (plugins_downloaded()) withAutoprint(\{ # examplesIf}
# Convert from platform name
client <- as_pjrt_client("cpu")
client
\dontshow{\}) # examplesIf}
}
