% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{as_pjrt_plugin}
\alias{as_pjrt_plugin}
\title{Convert to PJRT Plugin}
\usage{
as_pjrt_plugin(x)
}
\arguments{
\item{x}{(any)\cr
Object to convert to a PJRT plugin. Currently supports \code{PJRTPlugin} and \code{character(1)}.}
}
\value{
\code{PJRTPlugin}
}
\description{
Convert a platform name to a PJRT plugin or verify that an object is already a plugin.
}
\examples{
\dontshow{if (plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
# Convert from platform name
plugin <- as_pjrt_plugin("cpu")
plugin
\dontshow{\}) # examplesIf}
}
