% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{await}
\alias{await}
\title{Await an async operation}
\usage{
await(x, ...)
}
\arguments{
\item{x}{An async value object.}

\item{...}{Additional arguments (unused).}
}
\value{
The awaited object (invisibly).
}
\description{
Block until the async operation is complete and return the object.
}
