% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_from_device}
\alias{client_from_device}
\title{Get Client from Device}
\usage{
client_from_device(device)
}
\arguments{
\item{device}{(\code{PJRTDevice})\cr
A PJRT device object.}
}
\value{
\code{PJRTClient}
}
\description{
Get the client associated with a device by looking it up from the global cache.
}
\keyword{internal}
