% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{default_client}
\alias{default_client}
\title{Default Client}
\usage{
default_client()
}
\value{
\code{PJRTClient}
}
\description{
Respects environment variable \code{PJRT_DEVICE} and otherwise defaults to "cpu".
}
