% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_platform}
\alias{default_platform}
\title{Default Platform}
\usage{
default_platform()
}
\value{
\code{character(1)}
}
\description{
Respects environment variable \code{PJRT_PLATFORM} and otherwise defaults to "cpu".
}
