% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{device}
\alias{device}
\title{Device of a PJRTBuffer}
\usage{
device(x)
}
\arguments{
\item{x}{(any)\cr
Object for which to get the \code{PJRTDevice}.}
}
\value{
\code{PJRTDevice}
}
\description{
Get the device of a \code{\link[=pjrt_buffer]{PJRTBuffer}}.
}
