% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{devices}
\alias{devices}
\title{Devices}
\usage{
devices(x = NULL, ...)
}
\arguments{
\item{x}{An object to get devices from: a \code{\link[=pjrt_client]{PJRTClient}},
a \code{\link[=pjrt_compile]{PJRTLoadedExecutable}}, or \code{NULL} (default client).}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{list} of \code{PJRTDevice}
}
\description{
Get the addressable devices.
}
\examples{
\dontshow{if (plugins_downloaded()) withAutoprint(\{ # examplesIf}
# Create client (defaults to CPU)
client <- pjrt_client()
devices(client)
\dontshow{\}) # examplesIf}
}
