% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{dim.PJRTBuffer}
\alias{dim.PJRTBuffer}
\title{Dimenson of \code{PJRTBuffer}
Get the dimensions of a PJRT buffer}
\usage{
\method{dim}{PJRTBuffer}(x)
}
\arguments{
\item{x}{A PJRT buffer object.}
}
\value{
An integer vector of dimensions.
}
\description{
Dimenson of \code{PJRTBuffer}
Get the dimensions of a PJRT buffer
}
