% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{dtype}
\alias{dtype}
\title{Get the data type of a PJRTBuffer}
\usage{
dtype(x)
}
\arguments{
\item{x}{\code{\link[=pjrt_buffer]{PJRTBuffer}}}
}
\value{
\code{PJRTDtype}
}
\description{
Get the data type of a buffer, which includes its \link[=etype]{element type} and \link{shape}.
}
