% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{elt_type}
\alias{elt_type}
\title{Element Type}
\usage{
elt_type(x)
}
\arguments{
\item{x}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
Buffer.}
}
\description{
Get the element type of a buffer.
}
\examples{
\dontshow{if (plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
buf <- pjrt_buffer(c(1.0, 2.0, 3.0))
elt_type(buf)
\dontshow{\}) # examplesIf}
}
