% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{etype}
\alias{etype}
\title{Get the Element Type of a \code{PJRTBuffer}}
\usage{
etype(buffer)
}
\arguments{
\item{buffer}{A PJRT buffer object.}
}
\value{
A PJRT element type object.
}
\description{
Get the Element Type of a \code{PJRTBuffer}
}
