% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_buffer}
\alias{format_buffer}
\title{Format Buffer Data}
\usage{
format_buffer(buffer)
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
A PJRT buffer object.}
}
\value{
\code{character()} A character vector containing the formatted elements.
}
\description{
Formats buffer data into a character vector of string representations of individual elements suitable for StableHLO.
}
\examples{
\dontshow{if (plugins_downloaded()) withAutoprint(\{ # examplesIf}
buf <- pjrt_buffer(c(1.5, 2.5, 3.5))
format_buffer(buf)
\dontshow{\}) # examplesIf}
}
