% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{get_events}
\alias{get_events}
\title{Extract events from a buffer promise}
\usage{
get_events(x)
}
\arguments{
\item{x}{A buffer promise or other object}
}
\value{
A list of events
}
\description{
Extract all events from a buffer promise for chaining
}
\keyword{internal}
