% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{is_element_type}
\alias{is_element_type}
\title{Check if an object is a PJRT element type}
\usage{
is_element_type(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if the object is a PJRT element type, FALSE otherwise.
}
\description{
Check if an object is a PJRT element type
}
