% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pjrt_config}
\alias{pjrt_config}
\title{Configure PJRT}
\usage{
pjrt_config(cpu_device_count = NULL)
}
\arguments{
\item{cpu_device_count}{(\code{integer(1)})\cr
The number of CPU devices to use.}
}
\value{
(\code{list()})\cr
The current configuration.
}
\description{
Configure global PJRT options.
This needs to set before initializing the PJRT client, usually after loading the package.
}
