% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pjrt_cuda_versions}
\alias{pjrt_cuda_versions}
\title{Recommended CUDA Library Versions}
\format{
A named list with character elements:
\describe{
\item{cuda}{CUDA version}
\item{cudnn}{cuDNN version}
\item{nvshmem}{NVSHMEM version}
}
}
\usage{
pjrt_cuda_versions
}
\description{
A named list of the recommended library versions for CUDA support.
The PJRT CUDA plugin was built with these versions.
}
\keyword{datasets}
