% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{pjrt_device}
\alias{pjrt_device}
\title{Create a PJRT Device}
\usage{
pjrt_device(device)
}
\arguments{
\item{device}{(any)\cr
The device.}
}
\value{
\code{PJRTDevice}
}
\description{
Create a PJRT Device from an R object.
}
\section{Extractors}{

\itemize{
\item \code{\link[=platform]{platform()}} for a \code{character(1)} representation of the platform.
}
}

\examples{
\dontshow{if (plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
# Show available devices for CPU client
devices(pjrt_client("cpu"))
# Create device 0 for CPU client
dev <- pjrt_device("cpu:0")
dev
\dontshow{\}) # examplesIf}
}
