% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{pjrt_elt_type}
\alias{pjrt_elt_type}
\title{Get the data type of a PJRT buffer}
\usage{
pjrt_elt_type(buffer)
}
\arguments{
\item{buffer}{A PJRT buffer object.}
}
\value{
A PJRT element type object.
}
\description{
Get the data type of a PJRT buffer
}
