% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_buffer.R
\name{pjrt_lazy_buffer_materialize}
\alias{pjrt_lazy_buffer_materialize}
\title{Materialize a lazy buffer into a regular PJRTBuffer}
\usage{
pjrt_lazy_buffer_materialize(x)
}
\description{
Materialize a lazy buffer into a regular PJRTBuffer
}
