% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{pjrt_plugin}
\alias{pjrt_plugin}
\title{Create PJRT Plugin}
\usage{
pjrt_plugin(platform)
}
\arguments{
\item{platform}{(\code{character(1)})\cr
Platform name (e.g., "cpu", "cuda").}
}
\value{
\code{PJRTPlugin}
}
\description{
Create a PJRT plugin for a specific platform.
}
\section{Extractors}{

\itemize{
\item \code{\link[=plugin_attributes]{plugin_attributes()}} -> \code{list()}: for the attributes of the plugin.
}
}

\examples{
\dontshow{if (plugins_downloaded("cpu")) withAutoprint(\{ # examplesIf}
plugin <- pjrt_plugin("cpu")
plugin
\dontshow{\}) # examplesIf}
}
