% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{pjrt_plugin_attributes}
\alias{pjrt_plugin_attributes}
\title{Get Plugin Attributes}
\usage{
pjrt_plugin_attributes(plugin)
}
\arguments{
\item{plugin}{(\code{PJRTPlugin} | \code{character(1)})\cr
The plugin (or platform name) to get the attributes of.}
}
\value{
named \code{list()}
}
\description{
Get the attributes of a PJRT plugin.
This commonly includes:
\itemize{
\item \code{xla_version}
\item \code{stablehlo_current_version}
\item \code{stablehlo_minimum_version}
}

But the implementation depends on the plugin.
}
