% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{pjrt_program}
\alias{pjrt_program}
\title{Create a PJRTProgram}
\usage{
pjrt_program(code, format = c("mlir", "hlo"))
}
\arguments{
\item{code}{(\code{character(1)})\cr
Source code.}

\item{format}{(\code{character(1)})\cr
The format of the program.
One of "mlir" or "hlo".}
}
\value{
\code{\link{PJRTProgram}}
}
\description{
Create a program from a string.
}
