% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{pjrt_scalar}
\alias{pjrt_scalar}
\title{Create a PJRT scalar buffer}
\usage{
pjrt_scalar(data, client, ...)
}
\arguments{
\item{data}{A scalar value to convert to a PJRT buffer.}

\item{client}{A PJRT client object.}

\item{...}{Additional arguments.}
}
\value{
A PJRT buffer object representing a scalar.
}
\description{
Create a PJRT scalar buffer
}
