% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{platform}
\alias{platform}
\title{Platform Name}
\usage{
platform(x, ...)
}
\arguments{
\item{x}{(\code{PJRTBuffer})\cr
The buffer.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{character(1)}
}
\description{
Get the platform name of a PJRT buffer.
}
