% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{platform_name}
\alias{platform_name}
\title{Get the platform name of a PJRT client}
\usage{
platform_name(client = pjrt_client())
}
\arguments{
\item{client}{A PJRT client object.}
}
\value{
A string representing the platform name.
}
\description{
Get the platform name of a PJRT client
}
