% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_client_create}
\alias{plugin_client_create}
\title{Create PJRT Client}
\usage{
plugin_client_create(plugin, platform, options = list())
}
\arguments{
\item{plugin}{(\code{PJRTPlugin})\cr
The plugin to create a client for.}

\item{platform}{(\code{character(1)})\cr
The platform to create a client for.}

\item{options}{(\code{list()})\cr
Additional options to pass to the client.}
}
\value{
\code{PJRTClient}
}
\description{
Create a PJRT client for a specific plugin and platform.
}
