% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugin_is_downloaded}
\alias{plugin_is_downloaded}
\title{Check if Plugin is Downloaded}
\usage{
plugin_is_downloaded(platform = NULL)
}
\arguments{
\item{platform}{(\code{character(1)})\cr
Platform name.}
}
\value{
\code{logical(1)}
}
\description{
Check if a plugin is downloaded.
}
\examples{
\dontshow{if (plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
# Check if CPU plugin is downloaded
plugin_is_downloaded("cpu")
\dontshow{\}) # examplesIf}
}
