% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{plugins_downloaded}
\alias{plugins_downloaded}
\title{Check if Plugin is Downloaded}
\usage{
plugins_downloaded(platforms = NULL)
}
\arguments{
\item{platforms}{(\code{character()})\cr
Platform names.}
}
\value{
\code{logical(1)}
}
\description{
Check if one more more plugin is already downloaded.
}
\examples{
\dontshow{if (plugins_downloaded("cpu")) withAutoprint(\{ # examplesIf}
# Check if CPU plugin is downloaded
plugins_downloaded("cpu")
\dontshow{\}) # examplesIf}
}
