% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{shape}
\alias{dtype}
\alias{as_array}
\alias{as_raw}
\alias{device}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{xlamisc}{\code{\link[xlamisc:as_array]{as_array()}}, \code{\link[xlamisc:as_raw]{as_raw()}}, \code{\link[xlamisc:device]{device()}}, \code{\link[xlamisc:dtype]{dtype()}}, \code{\link[xlamisc:shape]{shape()}}}
}}

