% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{stablehlo_opt_available}
\alias{stablehlo_opt_available}
\title{Check if \code{stablehlo-opt} is Downloaded}
\usage{
stablehlo_opt_available()
}
\value{
\code{logical(1)}
}
\description{
Whether the \code{stablehlo-opt} binary is available locally, i.e. whether
\code{\link[=pjrt_optimize]{pjrt_optimize()}} can be used without a download.
}
\examples{
stablehlo_opt_available()
}
