% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as_array}
\alias{as_array}
\title{Convert a PJRT Buffer to an R object.}
\usage{
as_array(buffer, client = pjrt_client())
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
A PJRT buffer object.}

\item{client}{(\code{PJRTClient} | \code{character(1)})\cr
A PJRT client object or the name of the platform to use ("cpu", "cuda", ...),
from which the client will be created.}
}
\description{
Copy a \code{\link[=pjrt_buffer]{PJRTBuffer}} to an R object.
For 0-dimensional PJRT buffers, the R object will be a vector of length 1 and otherwise an array.
}
\details{
Moving the buffer to the host requires to:
\enumerate{
\item Copy the data from the PJRT device to the CPU.
\item Transpose the data, because PJRT returns it in row-major order but R uses column-major order.
}
}
