% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as_raw}
\alias{as_raw}
\title{Convert a PJRT Buffer to a raw R vector.}
\usage{
as_raw(buffer, client = default_client(), row_major)
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
A PJRT buffer object.}

\item{client}{(\code{PJRTClient} | \code{character(1)})\cr
A PJRT client object or the name of the platform to use ("cpu", "cuda", ...),
from which the client will be created.}

\item{row_major}{(\code{logical(1)})\cr
Whether to return the data in row-major format (TRUE) or column-major format (FALSE).
R uses column-major format.}
}
\value{
\code{raw()}
}
\description{
Copy a \code{\link[=pjrt_buffer]{PJRTBuffer}} to a raw R vector containing the buffer data as bytes.
Any shape information is lost.
}
