% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{devices}
\alias{devices}
\title{Devices}
\usage{
devices(client = NULL)
}
\arguments{
\item{client}{(\code{\link[=pjrt_client]{PJRTClient}})\cr
Object convertible to a \code{PJRTClient}.}
}
\value{
\code{list} of \code{PJRTDevice}
}
\description{
Get the addressable devices for a PJRT client.
}
