% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{dtype}
\alias{dtype}
\title{Data Type of Buffer}
\usage{
dtype(buffer)
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
Buffer.}
}
\value{
A PJRT element type object.
}
\description{
Data Type of Buffer
}
