% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{dtype}
\alias{dtype}
\title{Get the data type of a PJRTBuffer}
\usage{
dtype(x)
}
\arguments{
\item{x}{\code{\link[=pjrt_buffer]{PJRTBuffer}}}
}
\value{
A \code{PJRTDtype} object.
}
\description{
Get the data type of a PJRTBuffer.
}
