% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{flatten_fun}
\alias{flatten_fun}
\title{Flatten a Function}
\usage{
flatten_fun(f, in_tree)
}
\arguments{
\item{f}{(\code{function})\cr
Function to wrap.}

\item{in_tree}{(\code{RTree})\cr
Tree describing the (possibly nested) argument list of \code{f}, i.e.
\code{build_tree(list(arg1, arg2, ...))}. The flat inputs to the wrapper follow
this structure.}
}
\value{
A \code{FlattenedFunction}.
}
\description{
Flattens a function taking in possibly nested \code{RTree} objects so it takes in flat inputs
and return flat outputs (along the returned tree structure).
}
\examples{
f <- function(x) list(sum = x$a + x$b)
ff <- flatten_fun(f, build_tree(list(list(a = 1, b = 2))))
ff(3, 4)
}
\seealso{
\code{\link[=build_tree]{build_tree()}}, \code{\link[=flatten]{flatten()}}, \code{\link[=unflatten]{unflatten()}}
}
