% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjrt-package.R
\docType{package}
\name{pjrt-package}
\alias{pjrt}
\alias{pjrt-package}
\title{pjrt: R Interface to PJRT}
\description{
Provides an R interface to PJRT (Pretty much Just another RunTime), which allows you to run XLA or stableHLO programs on a variety of hardware backends including CPU, GPU, and TPU.
}
\section{Environment Variables}{

\itemize{
\item \code{TF_CPP_MIN_LOG_LEVEL}: Logging level for PJRT C++ API:
\itemize{
\item 0: shows info, warnings and errors
\item 1: shows warnings and errors
\item 2: shows errors
\item 3: shows nothing
}
\item \code{PJRT_PLATFORM}: Default platform to use, falls back to \code{"cpu"}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://r-xla.github.io/pjrt/}
  \item \url{https://github.com/r-xla/pjrt}
  \item Report bugs at \url{https://github.com/r-xla/pjrt/issues}
}

}
\author{
\strong{Maintainer}: Daniel Falbel \email{daniel@posit.co} (\href{https://orcid.org/0009-0006-0143-2392}{ORCID})

Authors:
\itemize{
  \item Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})
}

}
