% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjrt-package.R
\docType{package}
\name{pjrt-package}
\alias{pjrt}
\alias{pjrt-package}
\title{pjrt: R Interface to PJRT}
\description{
Provides an R interface to PJRT (Pretty much Just another RunTime), which allows you to run XLA or stableHLO programs on a variety of hardware backends including CPU, GPU, and TPU.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-xla.github.io/pjrt/}
  \item \url{https://github.com/r-xla/pjrt}
  \item Report bugs at \url{https://github.com/r-xla/pjrt/issues}
}

}
\author{
\strong{Maintainer}: Daniel Falbel \email{daniel@posit.co} (\href{https://orcid.org/0009-0006-0143-2392}{ORCID})

Authors:
\itemize{
  \item Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})
}

}
\keyword{internal}
