% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pjrt-package}
\alias{pjrt}
\alias{pjrt-package}
\title{pjrt: R Interface to PJRT}
\description{
Provides an R interface to PJRT (Pretty much Just another RunTime), which allows you to run XLA or stableHLO programs on a variety of hardware backends including CPU, GPU, and TPU.
}
\section{Environment Variables}{

\strong{Configuration options provided by XLA}

XLA provides various configuration options, but their documentation is scattered across
various websites. The options include:
\itemize{
\item \code{TF_CPP_MIN_LOG_LEVEL}: Logging level for PJRT C++ API:
\itemize{
\item 0: shows info, warnings and errors
\item 1: shows warnings and errors
\item 2: shows errors
\item 3: shows nothing
}
\item \code{XLA_FLAGS}: See the \href{https://openxla.org/xla/flags_guidance}{openxla website} for
more information.
}

\strong{Configuration options provided by this package}
\itemize{
\item \code{PJRT_PLATFORM}: Default platform to use, falls back to \code{"cpu"}.
\item \verb{PJRT_PLUGIN_PATH_<PLATFORM>}: Path to custom plugin library file for a specific
platform (e.g., \code{PJRT_PLUGIN_PATH_CPU}, \code{PJRT_PLUGIN_PATH_CUDA},
\code{PJRT_PLUGIN_PATH_METAL}). If set, the package will use this path instead
of downloading the plugin.
\item \verb{PJRT_PLUGIN_URL_<PLATFORM>}: URL to download plugin from for a specific
platform (e.g., \code{PJRT_PLUGIN_URL_CPU}, \code{PJRT_PLUGIN_URL_CUDA},
\code{PJRT_PLUGIN_URL_METAL}). If set, overrides the default plugin download URL.
\item \code{PJRT_ZML_ARTIFACT_VERSION}: Version of ZML artifacts to download.
Only used when downloading plugins from zml/pjrt-artifacts.
\item \code{PJRT_CPU_DEVICE_COUNT}: The number of CPU devices to use. Defaults to 1.
This is primarily intended for testing purposes.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://r-xla.github.io/pjrt/}
  \item \url{https://github.com/r-xla/pjrt}
  \item Report bugs at \url{https://github.com/r-xla/pjrt/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})

Authors:
\itemize{
  \item Daniel Falbel \email{daniel@posit.co} (\href{https://orcid.org/0009-0006-0143-2392}{ORCID})
}

}
