% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{pjrt_buffer}
\alias{pjrt_buffer}
\alias{pjrt_scalar}
\alias{pjrt_buffer.logical}
\alias{pjrt_buffer.integer}
\alias{pjrt_buffer.double}
\alias{pjrt_buffer.raw}
\alias{pjrt_scalar.logical}
\alias{pjrt_scalar.integer}
\alias{pjrt_scalar.double}
\alias{pjrt_scalar.raw}
\title{Create a PJRT Buffer}
\usage{
pjrt_buffer(data, elt_type, client = pjrt_client(), ...)

pjrt_scalar(data, elt_type, client = pjrt_client(), ...)

\method{pjrt_buffer}{logical}(
  data,
  elt_type = "pred",
  client = pjrt_client(),
  dims = get_dims(data),
  ...
)

\method{pjrt_buffer}{integer}(
  data,
  elt_type = "i32",
  client = pjrt_client(),
  dims = get_dims(data),
  ...
)

\method{pjrt_buffer}{double}(
  data,
  elt_type = "f32",
  client = pjrt_client(),
  dims = get_dims(data),
  ...
)

\method{pjrt_buffer}{raw}(data, ..., elt_type, client = pjrt_client(), shape, row_major)

\method{pjrt_scalar}{logical}(data, elt_type = "pred", client = pjrt_client(), ...)

\method{pjrt_scalar}{integer}(data, elt_type = "i32", client = pjrt_client(), ...)

\method{pjrt_scalar}{double}(data, elt_type = "f32", client = pjrt_client(), ...)

\method{pjrt_scalar}{raw}(data, ..., elt_type, client = pjrt_client())
}
\arguments{
\item{data}{(any)\cr
Data to convert to a \code{PJRTBuffer}.}

\item{elt_type}{(\code{character(1)})\cr
The type of the buffer.
Currently supported types are:
\itemize{
\item \code{"pred"}: predicate (i.e. a boolean)
\item \code{"{s,u}{8,16,32,64}"}: (Un)signed integer (for \code{integer} data).
\item \code{"f{32,64}"}: Floating point (for \code{double} data).
}}

\item{client}{(\code{PJRTClient} | \code{character(1)})\cr
A PJRT client object or the name of the platform to use ("cpu", "cuda", ...),
from which the client will be created.}

\item{...}{(any)\cr
Additional arguments.}

\item{dims}{(integer)\cr
The dimensions of the buffer.
If \code{NULL}, the dimensions will be inferred from the data.}
}
\value{
\code{PJRTBuffer}
}
\description{
Create a PJRT Buffer from an R object.
Any numeric PJRT buffer is an array and 0-dimensional arrays are used as scalars.
\code{\link{pjrt_buffer}} will create a array with dimensions \code{(1)} for a vector of length 1, while
Therefore, \code{\link{pjrt_scalar}} will create a 0-dimensional array for an R vector of length 1.
}
\details{
R does not have 0-dimensional arrays, hence we need the extra \code{pjrt_scalar} function.
}
\section{Extractors}{

\itemize{
\item \code{\link[=device]{device()}} for the device of the buffer.
\item \code{\link[=element_type]{element_type()}} for the element type of the buffer.
}
}

\section{Converters}{

\itemize{
\item \code{\link[=as_array]{as_array()}} for an array.
\item \code{\link[=as_raw]{as_raw()}} for a raw vector.
}
}

