% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{pjrt_buffer}
\alias{pjrt_buffer}
\title{Create a PJRT buffer}
\usage{
pjrt_buffer(data, client = default_client(), ...)
}
\arguments{
\item{data}{A vector of data to convert to a PJRT buffer.}

\item{client}{A PJRT client object.}

\item{...}{Additional arguments.}
}
\value{
A PJRT buffer object.
}
\description{
Create a PJRT buffer
}
