% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pjrt_compile}
\alias{pjrt_compile}
\title{Compile a Program}
\usage{
pjrt_compile(
  program,
  compile_options = new_compile_options(),
  client = pjrt_client()
)
}
\arguments{
\item{program}{(\code{character(1)})\cr
A program to compile.}

\item{compile_options}{(\code{PJRTCompileOptions})\cr
Compile options.}

\item{client}{(\code{PJRTClient} | \code{character(1)})\cr
A PJRT client object or the name of the platform to use ("cpu", "cuda", ...),
from which the client will be created.}
}
\value{
\code{PJRTExecutable}
}
\description{
Compile a \code{PJRTProgram} program into a \code{PJRTExecutable}.
}
\examples{
\dontshow{if (plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
# Create a simple program
src <- r"(
func.func @main(\%arg0: tensor<2xf32>) -> tensor<2xf32> {
  return \%arg0 : tensor<2xf32>
}
)"
prog <- pjrt_program(src = src)
exec <- pjrt_compile(prog)
\dontshow{\}) # examplesIf}
}
