% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pjrt_compile}
\alias{pjrt_compile}
\title{Compile a Program}
\usage{
pjrt_compile(
  program,
  compile_options = new_compile_options(),
  client = default_client()
)
}
\arguments{
\item{program}{(\code{character(1)})\cr
A program to compile.}

\item{compile_options}{(\code{\link{PJRTCompileOptions}})\cr
Compile options.}

\item{client}{(\code{\link{PJRTClient}})\cr
A PJRT client object.}
}
\value{
\code{\link{PJRTExecutable}}
}
\description{
Compile a \code{\link{PJRTProgram}} program into a \code{\link{PJRTExecutable}}.
}
