% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{pjrt_compile}
\alias{pjrt_compile}
\title{Compile a Program}
\usage{
pjrt_compile(
  program,
  compile_options = new_compile_options(),
  client = pjrt_client("cpu")
)
}
\arguments{
\item{program}{(\code{character(1)})\cr
A program to compile.}

\item{compile_options}{(\code{PJRTCompileOptions})\cr
Compile options.}

\item{client}{(\code{PJRTClient} | \code{character(1)})\cr
A PJRT client object or the name of the device to use, from which the client will be created.}
}
\value{
\code{PJRTExecutable}
}
\description{
Compile a \code{PJRTProgram} program into a \code{PJRTExecutable}.
}
