% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{pjrt_device}
\alias{pjrt_device}
\title{Device of a PJRTBuffer}
\usage{
pjrt_device(buffer)
}
\arguments{
\item{buffer}{(\code{\link[=pjrt_buffer]{PJRTBuffer}})\cr
A PJRT buffer object.}
}
\value{
\code{PJRTDevice}
}
\description{
Get the device of a \code{\link[=pjrt_buffer]{PJRTBuffer}}.
}
