% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_executable.R
\name{pjrt_execute}
\alias{pjrt_execute}
\title{Execute a PJRT program}
\usage{
pjrt_execute(executable, ...)
}
\arguments{
\item{executable}{(\code{PJRTLoadedExecutable})\cr
A PJRT program.}

\item{...}{Inputs to the program.}
}
\value{
A list of \code{PJRTBuffer}s.
}
\description{
Execute a PJRT program with the given inputs.
}
