% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_executable.R
\name{pjrt_execute}
\alias{pjrt_execute}
\title{Execute a PJRT program}
\usage{
pjrt_execute(executable, ..., execution_options = NULL)
}
\arguments{
\item{executable}{(\code{PJRTLoadedExecutable})\cr
A PJRT program.}

\item{...}{(\verb{PJRTBuffer)}\cr
Inputs to the program.
Named are ignored and arguments are passed in order.}

\item{execution_options}{(\code{PJRTExecuteOptions})\cr
Optional execution options for configuring buffer donation and other settings.}
}
\value{
\code{PJRTBuffer} | \code{list} of \code{PJRTBuffer}s
}
\description{
Execute a PJRT program with the given inputs and execution options.

\strong{Important:}
Arguments are passed by position and names are ignored.
}
