% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{pjrt_plugin}
\alias{pjrt_plugin}
\title{Create PJRT Plugin}
\usage{
pjrt_plugin(platform)
}
\arguments{
\item{platform}{(\code{character(1)})\cr
Platform name (e.g., "cpu", "cuda", "metal").}
}
\value{
\code{PJRTPlugin}
}
\description{
Create a PJRT plugin for a specific platform.
}
