% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{pjrt_program}
\alias{pjrt_program}
\title{Create a \code{PJRTProgram}}
\usage{
pjrt_program(src = NULL, path = NULL, format = c("mlir", "hlo"))
}
\arguments{
\item{src}{(\code{character(1)})
Source code.}

\item{path}{(\code{character(1)})
Path to the program file.}

\item{format}{(\code{character(1)})
One of "mlir" or "hlo".}
}
\value{
\code{PJRTProgram}
}
\description{
Create a program from a string or file path.
}
