% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{print.PJRTBuffer}
\alias{print.PJRTBuffer}
\title{Print a PJRT Buffer}
\usage{
\method{print}{PJRTBuffer}(
  x,
  max_rows = getOption("pjrt.print_max_rows", 30L),
  max_width = getOption("pjrt.print_max_width", 85L),
  max_rows_slice = getOption("pjrt.print_max_rows_slice", max_rows),
  header = TRUE,
  ...
)
}
\arguments{
\item{x}{(\code{PJRTBuffer})\cr
The buffer.}

\item{max_rows}{(\code{integer(1)})\cr
The maximum number of rows to print, excluding header and footer.}

\item{max_width}{(\code{integer(1)})\cr
The maximum width (in characters) of the printed buffer.
Set to negative values for no limit.
Note that for very small values, the actual printed width might be slightly smaller
as at least one column will be printed.
Also, this limit only affects the printed rows containing the actual data,
other rows might exceed the width.}

\item{max_rows_slice}{(\code{integer(1)})\cr
The maximum number of rows to print for each slice.}

\item{header}{(\code{logical(1)})\cr
Whether to print the header.}

\item{...}{Additional arguments (unused).}
}
\description{
Print a \code{\link[=pjrt_buffer]{PJRTBuffer}}.
}
